Consolidate file utils
