+++ collect_app/src/test/java/org/odk/collect/android/utilities/FileUtilsTest.java
@@ -1,5 +1,6 @@
 package org.odk.collect.android.utilities;
 
+import org.hamcrest.Matchers;
 import org.junit.Test;
 
 import java.io.BufferedWriter;
@@ -11,7 +12,10 @@ import java.util.HashMap;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.empty;
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class FileUtilsTest {
 
@@ -299,5 +303,17 @@ public class FileUtilsTest {
         assertThat(FileUtils.simplifyScopedStoragePath("blahblahblah"), is("blahblahblah"));
         assertThat(FileUtils.simplifyScopedStoragePath("/storage/emulated/0/Android/data/org.odk.collect.android/files/layers"), is("/sdcard/Android/data/org.odk.collect.android/files/layers"));
         assertThat(FileUtils.simplifyScopedStoragePath("/storage/emulated/0/Android/data/org.odk.collect.android/files/layers/countries/countries-raster.mbtiles"), is("/sdcard/Android/data/org.odk.collect.android/files/layers/countries/countries-raster.mbtiles"));
+    }
+
+    @Test
+    public void whenTryToListFilesOnNullFile_shouldReturnEmptyArray() {
+        assertThat(FileUtils.listFiles(null), Matchers.is(empty()));
+    }
+
+    @Test
+    public void whenTryToListFilesOnFileThatDoesNotExist_shouldReturnEmptyArray() {
+        File file = mock(File.class);
+        when(file.exists()).thenReturn(false);
+        assertThat(FileUtils.listFiles(file), Matchers.is(empty()));
     }
 }
